requirements.txt
asgiref==3.8.1
Django==5.1.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
sqlparse==0.5.1
tzdata==2024.2
